[RENAME] Star icon 파일 이동 #NADA-52
