+++ src/containers/activity/Thumbnail.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { Star } from '../../components/activity/Star';
+import { Star } from '../../components/common/icon/Star';
 import { caption_01, title_01 } from '../../styles/fontStyle';
 import HashTag from '../../components/activity/HashTag';
